Fix warnings

+++ Baikal/Utils/cl_program.cpp
@@ -192,8 +192,8 @@ CLWProgram CLProgram::Compile(const std::string &opts)
     }
     catch (CLWException exception)
     {
-        std::cerr<<"Compilation failed!"<<std::endl;
-        std::cerr<<"Dumping source to file:"<<m_program_name<<".cl.failed"<<std::endl;
+        std::cerr << "Compilation failed!" << std::endl;
+        std::cerr << "Dumping source to file:" << m_program_name << ".cl.failed" << std::endl;
         std::string fname = m_program_name + ".cl.failed";
         std::ofstream file(fname);
         file << m_compiled_source;
@@ -202,8 +202,8 @@ CLWProgram CLProgram::Compile(const std::string &opts)
     }
 
     end = std::chrono::high_resolution_clock::now();
-    int elapsed_ms = std::chrono::duration_cast<std::chrono::milliseconds>(end-start).count();
-    std::cerr<<"Program compilation time: "<<elapsed_ms<<" ms"<<std::endl;
+    int elapsed_ms = (int)std::chrono::duration_cast<std::chrono::milliseconds>(end - start).count();
+    std::cerr << "Program compilation time: " << elapsed_ms << " ms" << std::endl;
 
     m_is_dirty = false;
     return compiled_program;
